bug fixing. I don't exactly know the problem but it is very closed to be solved. the critical point is shifting as M increases for explosive percolation

+++ src/nets/network.h
@@ -78,7 +78,7 @@ public:
         return ss.str();
     }
 
-    std::vector<double> degrees();
+    std::vector<size_t> degrees();
     double clusteringCoefficient() const {
         cout << "TODO : line " << __LINE__ << endl;
         return 0;
